Keyfiles should be exactly KEYFILE_SIZE bytes long.
Indeed, when reading a keyfile, we should have read exactly KEYFILE_SIZE
bytes, not less. Reading less means that the file is truncated and we
should not bother with the rest of the processing (checking whether it
is protected, decrypting, copying, etc.). We should just bail out with
an ERR_READ error.