FIX: Bug when the user attempts to edit a keyfile.
This happens when the user starts editing a keyfile but doesn't give the
same passphrase for confirmation he/she gave when asked for it.
The bug lies in that gen_key() fails because the called get_passphrase()
fails, thus write_key() bails out with a keyfile open in 'wb'
(thus truncated) without writing anything into it.