New version of key files format with Key Derivation params
This version addresses issue #3. Indeed, instead of hard-coding password
hash parameters in the program itself, we give the opportunity to the
user to specify them in the command line. Now, some defaults of these
parameters are actually present in the code and are used if the user
doesn't specify them.